PrefixSelector refactor
