+++ src/utils/prefixUtils.js
@@ -1,15 +1,26 @@
+import Codemirror from 'codemirror';
 import Editor from '../entities/editor';
 import Prefix from '../entities/shexEntities/shexUtils/prefix.js';
 
 export function getPrefix(prefix){
-      let defined = Editor.getInstance().getYashe().getDefinedPrefixes();
-      for(let def in defined){
-          if(defined[def] == prefix){
-            return new Prefix(def,defined[def]);
-          }
-      }
-      return new Prefix();
+    let defined = Editor.getInstance().getYashe().getDefinedPrefixes();
+    for(let def in defined){
+        if(defined[def] == prefix){
+          return new Prefix(def,defined[def]);
+        }
+    }
+    return new Prefix();
 }
+
+export function addPrefix(prefix,namespaces){
+  let yashe = Editor.getInstance().getYashe();
+  let current = yashe.getValue();
+  let defined = yashe.getDefinedPrefixes();
+  let uri = getUri(prefix,namespaces);
+  yashe.setValue( 'PREFIX ' + prefix + ': <' + uri + '>\n' + current );
+        Codemirror.signal(yashe,'prefixUpdate');
+}
+
 
 export function getUri(prefix,namespaces){
   for(let def in namespaces){
@@ -18,4 +29,5 @@ export function getUri(prefix,namespaces){
         return namespaces[def][p];
     }
   }
+  return 'http://example.org/';
 }
